initial 2.0 setup
For the moment, I think removing ./lib/ is just not going to be
possible because we still need lots of cross-merges between
1.4 and 2.0.   Unfortunately I think we are stuck with it.

As for other 2.0 related things, we are still watching 1.4
move along and I don't think we are ready for the "official"
2.0 branch yet, as it will keep getting too far behind.
So for now, maintain one or more "experimental" 2.0 branches
where we can at least maintain commits we are going to need,
like adding the 2.0 migration and all that stuff.   But
don't name anything "rel_2_0" yet and these branches can
disappear at any time.

Change-Id: Ib5af75df94b23104eebe0e918adcf979d798ea3b